Create per-partition map, working
To create per-partition maps, we need to ensure that each map starts
mapping the output from position zero onwards. The only remaining
problem is that we get too many partitions, since also the logical
boot records and the empty spaces after them are mapped. However, this
is not a big deal for finding LVM partitions - LVM just does not care.